backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
mysqlclient==2.0.3
